Add .hushlogin with Ghostty.
Alacritty directly launch arguments passed to it, without
wrapping login. Ghostty however follows macOS conventions
and does the login wrapping (and I am debating if my hatred
of unnecessary files in the home directory is extreme enough
to warrant a PR just to add a config option). Thus the only way
to get rid of the "Last login time" on macOS is by adding
the .hushlogin file to the home directory.