firefox: move adblock and sponsorblock to behind boolean, disabled by default
I know that it is painful to use the internet w/o adblock, but I simply am not sure if it is ethical
to use them, as it sometimes seems like taking them for fools and running off with their product,
which when put that way sounds a lot like theft (Comm. 8) or reneging on an agreement (Comm. 9).
This was prompted by a friend asking the fundamental (not legal) difference between piracy and
adblock, which admittedly I had not thought about previously.