shell: revert to zsh for the shell
what started as a switch to fish over frustrations regarding nushell
just resulted in frustrations regarding fish as well, so back to
zsh we go. The issue with nushell was that it simply was too non-
posix complicant which was just a pain for a variety of reasons
(e.g. nix path env var being broken because of the paramter expansion
nix expected it could do). Yet with fish it wouldn't even source
path at all, so I could get locked out of nix (thankfully I left the
other shells installed in the mean time). So it's back to zsh I guess