emacs: (finally) fix the path variables
Okay, I think I finally have it fixed.
With stock exec-path-from-shell, the path additions
not found by the shell are dropped, so we need to
cache the nix locations prior to epfs initialization.
But I also have to update the PATH envvar as well, so
that envrc sees it in the process-environment and
keeps them, otherwise it just drops them from exec-path
as well when applying that.

And just for good measure don't init direnv before epfs.